Move other calibration code out of commander
