+++ src/modules/commander/CMakeLists.txt
@@ -41,7 +41,6 @@ px4_add_module(
 		commander.cpp
 		state_machine_helper.cpp
 		commander_helper.cpp
-		calibration_routines.cpp
 		baro_calibration.cpp
 		PreflightCheck.cpp
 		arm_auth.cpp
