+++ src/modules/events/calibration/accelerometer_calibration.cpp
@@ -123,8 +123,8 @@
 // FIXME: Can some of these headers move out with detect_ move?
 
 #include "accelerometer_calibration.h"
-#include <commander/calibration_messages.h>
-#include <commander/calibration_routines.h>
+#include "calibration_messages.h"
+#include "calibration_routines.h"
 
 #include <px4_defines.h>
 #include <px4_posix.h>
