+++ src/modules/events/calibration/calibration_routines.cpp
@@ -59,7 +59,7 @@
 
 #include "calibration_routines.h"
 #include "calibration_messages.h"
-#include "commander_helper.h"
+#include <commander/commander_helper.h>
 
 int sphere_fit_least_squares(const float x[], const float y[], const float z[],
 			     unsigned int size, unsigned int max_iterations, float delta, float *sphere_x, float *sphere_y, float *sphere_z,
