+++ src/modules/events/calibration/esc_calibration.cpp
@@ -40,8 +40,8 @@
  */
 
 #include "esc_calibration.h"
-#include <commander/calibration_messages.h>
-#include <commander/calibration_routines.h>
+#include "calibration_messages.h"
+#include "calibration_routines.h"
 
 #include <stdio.h>
 #include <stdlib.h>
