+++ src/modules/events/calibration/mag_calibration.cpp
@@ -38,8 +38,8 @@
  */
 
 #include "mag_calibration.h"
-#include <commander/calibration_routines.h>
-#include <commander/calibration_messages.h>
+#include "calibration_routines.h"
+#include "calibration_messages.h"
 
 #include <px4_defines.h>
 #include <px4_posix.h>
